+++ librt/oslo_calc.c
@@ -200,8 +200,9 @@ struct oslo_mat * om;
 
 	for( omp = om; omp!= ( struct oslo_mat *) 0; omp = omp->next)
 	{
-		fprintf( stderr, "%lx offset %d osize %d next %lx\n",  omp,
-		    omp->offset,  omp->osize,  omp->next);
+		fprintf(stderr, "%lx offset %d osize %d next %lx\n",
+			(unsigned long)omp,  omp->offset,  omp->osize,
+			(unsigned long)omp->next);
 
 		fprintf(stderr,"\t%f",  omp->o_vec[0]);
